+++ src/SixToFix.Application/GlobalUsings.cs
@@ -6,4 +6,6 @@ global using System.Threading.Tasks;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Logging;
 global using SixToFix.Domain;
+global using SixToFix.Domain.Constants;
 global using SixToFix.Domain.Entities;
+global using SixToFix.Domain.Enums;