+++ src/SixToFix.Domain/Constants/Roles.cs
@@ -1,0 +1,8 @@
+namespace SixToFix.Domain.Constants;
+
+public static class Roles
+{
+    public const string SuperAdmin = "SuperAdmin";
+    public const string TenantAdmin = "TenantAdmin";
+    public const string Client = "Client";
+}